completed code
Depending on the size of the graph, a blank "n x n" matrix is created. 

The user can manually create the graph by determining which vertices are created by using the "add_edge()" function.

When a graph is made, you can create all permutations of the graph with "create_permuations", along with the matrices of those permuations with "create_matrices()".

Finally, you can create the second graph, and use "compare_matrices()" to check if there are isomers.